CRUD complete

+++ public/scripts/item.js
@@ -1,23 +1,20 @@
 let initialName = document.getElementById("name").value;
 let initialPrice = document.getElementById("price").value;
 let initialStock = document.getElementById("stock").value;
+let initialDescription = document.getElementById("description").value;
 let tags = [];
-if(document.getElementById("tagsHidden").value)
-  tags = document.getElementById("tagsHidden").value.split(",");
-let warehouses = []
-if(document.getElementById("warehousesHidden").value)
-  warehouses = document.getElementById("warehousesHidden").value.split(",");
+let warehouses = [];
 
-let deleteItem = ()=>{
+let deleteItem = () => {
   let d = confirm("Are you sure you want to delete this item?");
-  if(d){
+  if (d) {
     fetch(location.pathname, {
-      method: "DELETE"
+      method: "DELETE",
     })
       .then((response) => {
         if (response.status === 204) {
-          alert('Deleted successfully!')
-          location.href="/items";
+          alert("Deleted successfully!");
+          location.href = "/items";
         } else {
           alert("An error occurred :(");
         }
@@ -26,53 +23,63 @@ let deleteItem = ()=>{
         console.error(err);
       });
   }
-}
+};
 
 let resetDefaults = () => {
   tags = [];
-  if(document.getElementById("tagsHidden").value)
+  if (document.getElementById("tagsHidden").value)
     tags = document.getElementById("tagsHidden").value.split(",");
-  warehouses = []
-  if(document.getElementById("warehousesHidden").value)
+  warehouses = [];
+  if (document.getElementById("warehousesHidden").value)
     warehouses = document.getElementById("warehousesHidden").value.split(",");
   document.getElementById("name").value = initialName;
   document.getElementById("stock").value = initialStock;
   document.getElementById("price").value = initialPrice;
+  document.getElementById("description").value = initialDescription;
   document.getElementById("tags").innerHTML = "";
   for (let tag of tags) {
     visualTagAdd(tag);
   }
   let wc = document.getElementsByClassName("warehouseCheck");
-  for(c of wc){
+  for (c of wc) {
     c.checked = false;
   }
-  for(let w of warehouses){
+  for (let w of warehouses) {
     document.getElementById(w).checked = true;
   }
 };
 
 let toggleEdit = (dis) => {
-  document.getElementById("edit").style.display = dis ? "inline" : "none";
+  if (document.getElementById("edit"))
+    document.getElementById("edit").style.display = dis ? "inline" : "none";
   document.getElementById("name").disabled = dis;
   document.getElementById("stock").disabled = dis;
   document.getElementById("price").disabled = dis;
+  document.getElementById("description").disabled = dis;
   document.getElementById("addingTags").style.display = dis ? "none" : "block";
   document.getElementById("wareLocs").style.display = dis ? "none" : "block";
   document.getElementById("saveBtn").style.display = dis ? "none" : "inline";
-  document.getElementById("cancelBtn").style.display = dis ? "none" : "inline";
+  if(document.getElementById("cancelBtn"))
+    document.getElementById("cancelBtn").style.display = dis ? "none" : "inline";
   let rt = document.getElementsByClassName("removeTag");
-  for(let r of rt){
-    r.style.display = dis?"none":"inline";
+  for (let r of rt) {
+    r.style.display = dis ? "none" : "inline";
   }
-  if(document.getElementById("warehouses"))
-    document.getElementById("warehouses").style.display = dis ? "block" : "none";
+  if (document.getElementById("warehouses"))
+    document.getElementById("warehouses").style.display = dis
+      ? "block"
+      : "none";
 };
 
-let saveItem = ()=>{
+let saveItem = (method, path) => {
   let errorsPresent = false;
   if (!document.getElementById("name").value) {
     errorsPresent = true;
     document.getElementById("nameError").style.display = "inline";
+  }
+  if (!document.getElementById("description").value) {
+    errorsPresent = true;
+    document.getElementById("descriptionError").style.display = "inline";
   }
   if (
     !document.getElementById("price").value ||
@@ -90,27 +97,35 @@ let saveItem = ()=>{
   }
   if (!errorsPresent) {
     let data = {
-      "name": document.getElementById("name").value,
-      "stock": document.getElementById("stock").value,
-      "price": document.getElementById("price").value,
-      "tags": tags,
-      "warehouseLocations":warehouses
+      name: document.getElementById("name").value,
+      stock: document.getElementById("stock").value,
+      description: document.getElementById("description").value,
+      price: document.getElementById("price").value,
+      tags: tags,
+      warehouseLocations: warehouses,
     };
     document.getElementById("nameError").style.display = "none";
+    document.getElementById("descriptionError").style.display = "none";
     document.getElementById("priceError").style.display = "none";
     document.getElementById("stockError").style.display = "none";
-    fetch(location.pathname, {
-      method: "PUT",
+    fetch(path, {
+      method: method,
       body: JSON.stringify(data),
       headers: {
         "Content-Type": "application/json",
       },
     })
       .then((response) => {
-        if (response.status == 200) {
+        if (response.status === 200) {
           location.reload();
-        } else {
-          alert("An error occurred :(");
+        } else if(response.status === 201) {
+          alert("Item successfully added!")
+          location.href = "/items";
+        }
+        else {
+          response.json().then((data) => {
+            alert(`An error occurred: ${data.message}`);
+          });
         }
       })
       .catch((err) => {
@@ -119,7 +134,7 @@ let saveItem = ()=>{
   }
 };
 
-let visualTagAdd = (v)=>{
+let visualTagAdd = (v) => {
   let tagSpan = document.createElement("span");
   tagSpan.classList.add("badge");
   tagSpan.classList.add("badge-pill");
@@ -131,18 +146,37 @@ let visualTagAdd = (v)=>{
   document.getElementById("tags").appendChild(tagSpan);
 };
 
-document.getElementById("edit").addEventListener("click", () => {
+if (document.getElementById("createItem")) {
   toggleEdit(false);
+} else {
+  if (document.getElementById("tagsHidden").value)
+    tags = document.getElementById("tagsHidden").value.split(",");
+  if (document.getElementById("warehousesHidden").value)
+    warehouses = document.getElementById("warehousesHidden").value.split(",");
+}
+
+if (document.getElementById("edit") && document.getElementById("deleteItem")) {
+  document.getElementById("edit").addEventListener("click", () => {
+    toggleEdit(false);
+  });
+
+  document.getElementById("deleteItem").addEventListener("click", deleteItem);
+
+  document.getElementById("cancelBtn").addEventListener("click", () => {
+    resetDefaults();
+    toggleEdit(true);
+  });
+}
+
+document.getElementById("saveBtn").addEventListener("click", ()=>{
+  // checking if creating item or editing item
+  if(document.getElementById("createItem")){
+    saveItem("POST", "/items");
+  }
+  else{
+    saveItem("PUT", location.pathname);
+  }
 });
-
-document.getElementById("deleteItem").addEventListener("click", deleteItem);
-
-document.getElementById("cancelBtn").addEventListener("click", () => {
-  resetDefaults();
-  toggleEdit(true);
-});
-
-document.getElementById("saveBtn").addEventListener("click", saveItem);
 
 document.getElementById("tagAdd").addEventListener("keyup", (e) => {
   if (e.key == "," || e.key == " ") {
@@ -155,21 +189,20 @@ document.getElementById("tagAdd").addEventListener("keyup", (e) => {
   }
 });
 
-document.getElementById("wareLocs").addEventListener("click", (e)=>{
-  if(e.target.tagName === "INPUT"){
+document.getElementById("wareLocs").addEventListener("click", (e) => {
+  if (e.target.tagName === "INPUT") {
     console.log(e.target);
-    if(e.target.checked){
+    if (e.target.checked) {
       warehouses.push(e.target.id);
-    }
-    else{
+    } else {
       warehouses.splice(e.target.id, 1);
     }
-    console.log(warehouses)
+    console.log(warehouses);
   }
 });
 
-document.getElementById("tags").addEventListener("click", (e)=>{
-  if(e.target.className === "removeTag"){
+document.getElementById("tags").addEventListener("click", (e) => {
+  if (e.target.className === "removeTag") {
     e.target.parentNode.parentNode.removeChild(e.target.parentNode);
   }
-})
+});
